supabase/requirements.txt
# Supabase Backend Dependencies
# pip install -r supabase/requirements.txt

supabase>=2.0.0
openai>=1.0.0
pyyaml>=6.0
